# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Callora API",
    "version": "1.0.0",
    "description": "API contract covering Callora backend billing, usage, and developer routes."
  },
  "paths": {
...
    "/api/usage": {
      "get": {
        "summary": "Retrieve user API usage and stats",
        "description": "Returns the authenticated user's API call events, aggregated stats, and time breakdown.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start of period (ISO-8601 string)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End of period (ISO-8601 string)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of usage events to return",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "apiId",
            "in": "query",
            "description": "Filter by specific API ID",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "in": "query",
            "description": "Aggregation bucket size",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "day",
                "week",
                "month"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage events and analytics retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                },
                "examples": {
                  "withEvents": {
                    "summary": "Typical response with two usage events",
                    "value": {
                      "events": [
                        {
                          "id": "evt_01hx9r2k3m4n5p6q7r8s9t0u",
                          "apiId": "api_weather_v1",
                          "endpoint": "/v1/weather/current",
                          "occurredAt": "2026-07-01T14:32:00.000Z",
                          "revenue": "1000000"
                        },
                        {
                          "id": "evt_01hx9r2k3m4n5p6q7r8s9t0v",
                          "apiId": "api_weather_v1",
                          "endpoint": "/v1/weather/forecast",
                          "occurredAt": "2026-07-01T14:45:00.000Z",
                          "revenue": "1000000"
                        }
                      ],
                      "stats": {
                        "totalCalls": 2,
                        "totalSpent": "2000000",
                        "breakdownByApi": [
                          {
                            "apiId": "api_weather_v1",
                            "calls": 2,
                            "revenue": "2000000"
                          }
                        ]
                      },
                      "period": {
                        "from": "2026-06-01T00:00:00.000Z",
                        "to": "2026-07-01T00:00:00.000Z"
                      },
                      "pagination": {
                        "limit": 20,
                        "offset": 0,
                        "hasMore": false
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t0w"
                    }
                  },
                  "withBuckets": {
                    "summary": "Response with daily call buckets (groupBy=day)",
                    "value": {
                      "events": [
                        {
                          "id": "evt_01hx9r2k3m4n5p6q7r8s9t0x",
                          "apiId": "api_maps_v2",
                          "endpoint": "/v2/geocode",
                          "occurredAt": "2026-07-10T09:15:00.000Z",
                          "revenue": "500000"
                        }
                      ],
                      "stats": {
                        "totalCalls": 1,
                        "totalSpent": "500000",
                        "breakdownByApi": [
                          {
                            "apiId": "api_maps_v2",
                            "calls": 1,
                            "revenue": "500000"
                          }
                        ],
                        "buckets": [
                          {
                            "period": "2026-07-10",
                            "calls": 1,
                            "revenue": "500000"
                          }
                        ]
                      },
                      "period": {
                        "from": "2026-07-10T00:00:00.000Z",
                        "to": "2026-07-11T00:00:00.000Z"
                      },
                      "pagination": {
                        "limit": 20,
                        "offset": 0,
                        "hasMore": false
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t0y"
                    }
                  },
                  "empty": {
                    "summary": "No events in the requested period",
                    "value": {
                      "events": [],
                      "stats": {
                        "totalCalls": 0,
                        "totalSpent": "0",
                        "breakdownByApi": []
                      },
                      "period": {
                        "from": "2026-06-01T00:00:00.000Z",
                        "to": "2026-07-01T00:00:00.000Z"
                      },
                      "pagination": {
                        "limit": 20,
                        "offset": 0,
                        "hasMore": false
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t0z"
                    }
                  },
                  "withCursorPagination": {
                    "summary": "Cursor-paginated response (more pages available)",
                    "value": {
                      "events": [
                        {
                          "id": "evt_01hx9r2k3m4n5p6q7r8s9t1a",
                          "apiId": "api_translate_v1",
                          "endpoint": "/v1/translate",
                          "occurredAt": "2026-07-20T11:00:00.000Z",
                          "revenue": "250000"
                        }
                      ],
                      "stats": {
                        "totalCalls": 1,
                        "totalSpent": "250000",
                        "breakdownByApi": [
                          {
                            "apiId": "api_translate_v1",
                            "calls": 1,
                            "revenue": "250000"
                          }
                        ]
                      },
                      "period": {
                        "from": "2026-07-01T00:00:00.000Z",
                        "to": "2026-07-31T00:00:00.000Z"
                      },
                      "pagination": {
                        "limit": 1,
                        "nextCursor": "eyJjcmVhdGVkX2F0IjoiMjAyNi0wNy0yMFQxMTowMDowMC4wMDBaIiwiaWQiOiJldnRfMDFoeDlyMmszbTRuNXA2cTdyOHM5dDF"
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1b"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "invalidDateRange": {
                    "summary": "from date is after to date",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "'from' date must be before or equal to 'to' date",
                        "details": [
                          {
                            "field": "from",
                            "message": "'from' date must be before or equal to 'to' date",
                            "code": "custom"
                          }
                        ]
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1c",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  },
                  "invalidGroupBy": {
                    "summary": "Invalid groupBy value",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "VALIDATION_ERROR",
                        "message": "Invalid enum value. Expected 'day' | 'week' | 'month'",
                        "details": [
                          {
                            "field": "groupBy",
                            "message": "Invalid enum value. Expected 'day' | 'week' | 'month'",
                            "code": "invalid_enum_value"
                          }
                        ]
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1d",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  },
                  "invalidCursor": {
                    "summary": "Malformed cursor value",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "BAD_REQUEST",
                        "message": "Invalid cursor format. Cursor must be base64 encoded (created_at, id).",
                        "details": []
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1e",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "missingToken": {
                    "summary": "No authorization token provided",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "UNAUTHORIZED",
                        "message": "Unauthorized"
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1f",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  },
                  "expiredToken": {
                    "summary": "JWT token has expired",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "TOKEN_EXPIRED",
                        "message": "Token expired"
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1g",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "internalError": {
                    "summary": "Unexpected server error",
                    "value": {
                      "success": false,
                      "error": {
                        "code": "INTERNAL_SERVER_ERROR",
                        "message": "Internal server error"
                      },
                      "requestId": "req-01hx9r2k3m4n5p6q7r8s9t1h",
                      "timestamp": "2026-07-01T10:00:00.000Z"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/usage/anomalies": {
      "get": {
        "summary": "List detected usage anomalies",
        "description": "Returns per-API daily usage anomalies for admin review. Usage events are aggregated to per-API daily call counts over the requested window, then each API's days are scored against that API's own baseline using a z-score test; days whose absolute z-score meets the threshold are flagged as spikes or drops. Requires admin authentication and the admin IP allowlist.",
        "security": [
          {
            "adminApiKey": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "description": "Start of analysis window (ISO-8601). Defaults to 30 days ago.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "description": "End of analysis window (ISO-8601). Defaults to now.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "threshold",
            "in": "query",
            "description": "Absolute z-score at/above which a day is flagged.",
            "required": false,
            "schema": {
              "type": "number",
              "minimum": 1,
              "maximum": 10,
              "default": 3
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of anomalies to return (most severe first).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "apiId",
            "in": "query",
            "description": "Restrict analysis to a single API ID.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detected anomalies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "anomalies": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "apiId": {
                                "type": "string"
                              },
                              "day": {
                                "type": "string",
                                "example": "2026-03-21"
                              },
                              "type": {
                                "type": "string",
                                "enum": [
                                  "spike",
                                  "drop"
                                ]
                              },
                              "calls": {
                                "type": "integer"
                              },
                              "revenue": {
                                "type": "string"
                              },
                              "baselineMean": {
                                "type": "number"
                              },
                              "stdDev": {
                                "type": "number"
                              },
                              "zScore": {
                                "type": "number"
                              }
                            }
                          }
                        },
                        "summary": {
                          "type": "object",
                          "properties": {
                            "window": {
                              "type": "object",
                              "properties": {
                                "from": {
                                  "type": "string",
                                  "format": "date-time"
                                },
                                "to": {
                                  "type": "string",
                                  "format": "date-time"
                                }
                              }
                            },
                            "threshold": {
                              "type": "number"
                            },
                            "minDataPoints": {
                              "type": "integer"
                            },
                            "seriesAnalyzed": {
                              "type": "integer"
                            },
                            "anomalyCount": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by admin IP allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/usage/{developerId}": {
      "get": {
        "summary": "Inspect a developer usage aggregate",
        "description": "Returns a redacted aggregate snapshot for a developer's live usageStore state. Requires admin authentication and the admin IP allowlist.",
        "security": [
          {
            "adminApiKey": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "developerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Usage aggregate snapshot retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUsageSnapshotResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized admin request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by admin IP allowlist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usage aggregate not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "adminApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-api-key"
      }
    },
    "schemas": {
...
      "UsageResponse": {
        "type": "object",
        "required": [
          "events",
          "stats",
          "period"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "apiId",
                "endpoint",
                "occurredAt",
                "revenue"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Unique identifier for the usage event."
                },
                "apiId": {
                  "type": "string",
                  "description": "Identifier of the API that was called."
                },
                "endpoint": {
                  "type": "string",
                  "description": "API endpoint path that was invoked."
                },
                "occurredAt": {
                  "type": "string",
                  "format": "date-time",
                  "description": "ISO 8601 timestamp when the call occurred."
                },
                "revenue": {
                  "type": "string",
                  "description": "Revenue generated by this event in smallest USDC units (string to avoid precision loss)."
                }
              }
            }
          },
          "stats": {
            "type": "object",
            "required": [
              "totalCalls",
              "totalSpent",
              "breakdownByApi"
            ],
            "properties": {
              "totalCalls": {
                "type": "integer",
                "description": "Total number of API calls in the period."
              },
              "totalSpent": {
                "type": "string",
                "description": "Total spend in smallest USDC units (string to avoid precision loss)."
              },
              "breakdownByApi": {
                "type": "array",
                "description": "Per-API aggregated call count and revenue.",
                "items": {
                  "type": "object",
                  "required": [
                    "apiId",
                    "calls",
                    "revenue"
                  ],
                  "properties": {
                    "apiId": {
                      "type": "string"
                    },
                    "calls": {
                      "type": "integer"
                    },
                    "revenue": {
                      "type": "string"
                    }
                  }
                }
              },
              "buckets": {
                "type": "array",
                "description": "Time-bucketed aggregation (present when groupBy is supplied).",
                "items": {
                  "type": "object",
                  "required": [
                    "period",
                    "calls",
                    "revenue"
                  ],
                  "properties": {
                    "period": {
                      "type": "string",
                      "description": "Bucket label (e.g. YYYY-MM-DD for groupBy=day)."
                    },
                    "calls": {
                      "type": "integer"
                    },
                    "revenue": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "period": {
            "type": "object",
            "required": [
              "from",
              "to"
            ],
            "description": "The effective time range covered by this response.",
            "properties": {
              "from": {
                "type": "string",
                "format": "date-time"
              },
              "to": {
                "type": "string",
                "format": "date-time"
              }
            }
          },
          "pagination": {
            "type": "object",
            "description": "Pagination metadata. Shape varies by pagination mode: offset pagination includes limit/offset/hasMore; cursor pagination includes limit/nextCursor.",
            "properties": {
              "limit": {
                "type": "integer",
                "description": "Page size used for this request."
              },
              "offset": {
                "type": "integer",
                "description": "Offset applied for this request (offset pagination only)."
              },
              "hasMore": {
                "type": "boolean",
                "description": "Whether additional pages exist."
              },
              "nextCursor": {
                "type": "string",
                "description": "Opaque base64 cursor to pass as the `cursor` query parameter for the next page (cursor pagination only)."
              }
            }
          },
          "requestId": {
            "type": "string",
            "description": "Correlation ID for this request, propagated from the X-Request-Id header."
          }
        }
      },
      "AdminUsageSnapshot": {
        "type": "object",
        "required": [
          "developerId",
          "totalEvents",
          "settledEvents",
          "unsettledEvents",
          "totalAmountUsdc",
          "settledAmountUsdc",
          "unsettledAmountUsdc",
          "apiCount",
          "endpointCount",
          "firstEventAt",
          "lastEventAt",
          "statusCodes"
        ],
        "properties": {
          "developerId": {
            "type": "string"
          },
          "totalEvents": {
            "type": "integer"
          },
          "settledEvents": {
            "type": "integer"
          },
          "unsettledEvents": {
            "type": "integer"
          },
          "totalAmountUsdc": {
            "type": "number"
          },
          "settledAmountUsdc": {
            "type": "number"
          },
          "unsettledAmountUsdc": {
            "type": "number"
          },
          "apiCount": {
            "type": "integer"
          },
          "endpointCount": {
            "type": "integer"
          },
          "firstEventAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "lastEventAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "statusCodes": {
            "type": "object",
            "additionalProperties": {
              "type": "integer"
            }
          }
        }
      },
      "AdminUsageSnapshotResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AdminUsageSnapshot"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
    }
  }
}